+++ src/index.ts
@@ -12,8 +12,13 @@ import * as noNoopanimationsmodule from "./rules/deprecated/no-noopanimationsmod
 import * as noPlatformbrowserdynamic from "./rules/deprecated/no-platformbrowserdynamic.js";
 import * as noRoutertestingmodule from "./rules/deprecated/no-routertestingmodule.js";
 import * as noSubscribeInComponentConstructor from "./rules/experimental/no-subscribe-in-component-constructor.js";
+import * as noCanactivateClass from "./rules/functional/no-canactivate-class.js";
+import * as noCanactivatechildClass from "./rules/functional/no-canactivatechild-class.js";
+import * as noCandeactivateClass from "./rules/functional/no-candeactivate-class.js";
+import * as noCanmatchClass from "./rules/functional/no-canmatch-class.js";
 import * as noHttpinterceptorClass from "./rules/functional/no-httpinterceptor-class.js";
 import * as noHttpinterceptorsToken from "./rules/functional/no-httpinterceptors-token.js";
+import * as noResolveClass from "./rules/functional/no-resolve-class.js";
 import * as noWithinterceptorsfromdi from "./rules/functional/no-withinterceptorsfromdi.js";
 import * as noHostbindingDecorator from "./rules/host-bindings/no-hostbinding-decorator.js";
 import * as noHostlistenerDecorator from "./rules/host-bindings/no-hostlistener-decorator.js";
@@ -124,6 +129,11 @@ const plugin = {
     [noServermodule.ruleName]: noServermodule.ruleDefinition,
     [noWithmoduleTesting.ruleName]: noWithmoduleTesting.ruleDefinition,
     // Functional
+    [noCanactivateClass.ruleName]: noCanactivateClass.ruleDefinition,
+    [noCanactivatechildClass.ruleName]: noCanactivatechildClass.ruleDefinition,
+    [noCandeactivateClass.ruleName]: noCandeactivateClass.ruleDefinition,
+    [noCanmatchClass.ruleName]: noCanmatchClass.ruleDefinition,
+    [noResolveClass.ruleName]: noResolveClass.ruleDefinition,
     [noHttpinterceptorClass.ruleName]: noHttpinterceptorClass.ruleDefinition,
     [noHttpinterceptorsToken.ruleName]: noHttpinterceptorsToken.ruleDefinition,
     [noWithinterceptorsfromdi.ruleName]: noWithinterceptorsfromdi.ruleDefinition,
@@ -222,6 +232,11 @@ const functional: ConfigObject = {
     [name]: plugin
   },
   rules: {
+    [`${name}/${noCanactivateClass.ruleName}`]: "error",
+    [`${name}/${noCanactivatechildClass.ruleName}`]: "error",
+    [`${name}/${noCandeactivateClass.ruleName}`]: "error",
+    [`${name}/${noCanmatchClass.ruleName}`]: "error",
+    [`${name}/${noResolveClass.ruleName}`]: "error",
     [`${name}/${noHttpinterceptorClass.ruleName}`]: "error",
     [`${name}/${noHttpinterceptorsToken.ruleName}`]: "error",
     [`${name}/${noWithinterceptorsfromdi.ruleName}`]: "error",